requirements.txt
python-telegram-bot==21.8
anthropic>=0.25.0
python-dateutil==2.8.2
sqlalchemy==2.0.35
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
apscheduler==3.10.4
